requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
python-dotenv==1.0.0
pytest==7.4.3
openai==1.7.2
tiktoken==0.5.2
flask==3.0.0
flask-cors==4.0.0
